# docs/api.yaml
openapi: 3.0.4
info:
  title: Detective API
  version: v1
paths:
  /api/v1/auth:
    post:
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
      responses:
        '200':
          description: Аутентификация прошла успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
        '401':
          description: Неверный логин или пароль
        '503':
          description: Внутренний сервис не доступен
    patch:
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequestDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequestDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequestDto'
      responses:
        '204':
          description: Пароль успешно изменен
        '400':
          description: Новый пароль слабый
        '401':
          description: Неверный логин или пароль
        '503':
          description: Внутренний сервис не доступен
...
components:
  schemas:
    ChangePasswordRequestDto:
      type: object
      properties:
        username:
          type: string
          nullable: true
        oldPassword:
          type: string
          nullable: true
        newPassword:
          type: string
          nullable: true
      additionalProperties: false
...
    LoginRequestDto:
      type: object
      properties:
        username:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
    LoginResponseDto:
      type: object
      properties:
        token:
          type: string
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      description: Please enter a valid token
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]